+++ libcdi/src/timebase.c
@@ -194,9 +194,9 @@ double julday_sub(int julday1, int secofday1, int julday2, int secofday2, int *d
 
   *secs = (int) sec_of_day;
 
-  seconds = *days*86400. + sec_of_day;
+  seconds = *days * 86400 + sec_of_day;
 
-  return (seconds);
+  return (double)seconds;
 }
